requirements.txt
torch>=1.7.1,<2
numpy>=1.19.5,<2
fire>=0.4.0,<0.5.0
torchvision>=0.10.1,<2
clip-anytorch>=2.3.1,<3
scikit-video
